% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcounty.R
\docType{data}
\name{ke_subcounty}
\alias{ke_subcounty}
\title{Kenya Sub-County Boundary Map}
\format{
A simple feature collection with 47 features (counties) and sub-counties and 8 fields:
\describe{
\item{geometry}{MULTIPOLYGON geometry in WGS 84 (EPSG:4326)}
}
}
\source{
Kenya Independent Electoral and Boundaries Commission (IEBC)
}
\usage{
ke_subcounty
}
\description{
An \code{sf} object containing multipolygon boundary data for Kenya's
47 counties with their subcounties, based on the IEBC administrative boundary definitions.
}
\keyword{datasets}
